+++ client/src/io/worxbend/gitea4s/http/GitTreeParams.scala
@@ -1,0 +1,10 @@
+package io.worxbend.gitea4s.http
+
+final case class GitTreeParams(
+    recursive: Option[Boolean] = None,
+    page: Option[Int] = None,
+    perPage: Option[Int] = None
+)
+
+object GitTreeParams:
+  val default: GitTreeParams = GitTreeParams()